requirements.txt
certifi==2019.6.16
joblib==0.13.2
langdetect==1.0.7
numpy==1.16.4
pandas==0.24.2
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
ruamel.appconfig==0.5.4
ruamel.std.argparse==0.8.1
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
stop-words==2018.7.23
stopwords==0.1.3
tqdm==4.32.2
yaml-1.3==0.1.0
